+++ tools/openmedia/svg/src/schema.rs
@@ -12,6 +12,8 @@ pub enum JsonElement {
         ry: Option<f64>,
         fill: Option<String>,
         stroke: Option<String>,
+        stroke_width: Option<f64>,
+        opacity: Option<f64>,
     },
     Circle {
         cx: f64,
@@ -19,6 +21,18 @@ pub enum JsonElement {
         r: f64,
         fill: Option<String>,
         stroke: Option<String>,
+        stroke_width: Option<f64>,
+        opacity: Option<f64>,
+    },
+    Line {
+        x1: f64,
+        y1: f64,
+        x2: f64,
+        y2: f64,
+        stroke: Option<String>,
+        stroke_width: Option<f64>,
+        stroke_linecap: Option<String>,
+        opacity: Option<f64>,
     },
     Text {
         x: f64,
@@ -27,5 +41,8 @@ pub enum JsonElement {
         fill: Option<String>,
         font_size: Option<f64>,
         font_family: Option<String>,
+        font_weight: Option<u16>,
+        text_anchor: Option<String>,
+        opacity: Option<f64>,
     },
 }